중간 머지 작업 (#47)
* 가짜 WantedService제작 후 기능 구현함

* 제작한 가짜 WantedService와 실제 WantedService합침

* WantedDao 정리함

* WantedDao 위치 수정함

* Wanted 매퍼 수정함, WantedsDao 수정함

* Wanted의 Service, Controller 수정함

* Wanted 테스트용 가상 컨트롤러, 서비스 제작완료함

* wanted Mapper수정함

* likes Mapper 수정함

* like mapper, service, controller 구현함

* like mapper, service, controller 수정함함

* subcrives mapper, dao, service 구현함

* Code(Region, skill, career, position) Dao, Mapper, Service 작업함

* 테스트파일 삭제함

* 단위 테스트 제작

* 마스터 병합함

* apply mapper연결 밑 service 구현함

* subcribes mapper 연결 및 service 구현, Dto 구현함

* WantedController 구현 중, LikesService 추가함

* Apply, Wanted 서비스, 컨트롤러 제작함

* Dto static 수정함

* 마스터 pull 함

* wanted 컨트롤러 작업함

* Wanted 컨트롤러 제작함

* Wanted, Company, like Serviece 예외처리 제외 구현함

* wanted, company 컨트롤러 구현 완료

+++ src/main/java/site/metacoding/miniproject2/service/CompanysService.java
@@ -3,7 +3,10 @@ package site.metacoding.miniproject2.service;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 import site.metacoding.miniproject2.domain.companys.CompanysDao;
@@ -20,6 +23,7 @@ import site.metacoding.miniproject2.dto.CompanysRespDto.CompanysInsertRespDto;
 import site.metacoding.miniproject2.dto.CompanysRespDto.CompanysNumberCheckRespDto;
 import site.metacoding.miniproject2.dto.CompanysRespDto.SubscribesListRespDto;
 import site.metacoding.miniproject2.dto.SubribesReqDto.SubcribesInsertReqDto;
+import site.metacoding.miniproject2.dto.SubribesRespDto.SubcribesInsertRespDto;
 import site.metacoding.miniproject2.dto.SubribesRespDto.SubribesFindByIdRespDto;
 
 @RequiredArgsConstructor
@@ -29,7 +33,7 @@ public class CompanysService {
     private final SubcribesDao subcribesDao;
     private final CompanysDao companysDao;
     private final WantedsDao wantedsDao;
-    private final WantedsService wantedsService; // 쌤이 서비스에서 서비스 쓰지 말랬다고 알려주기
+    private final HttpSession session;
 
     /* 지원 작업 시작 */
     public CompanysInsertRespDto insertCompany(CompanysInsertReqDto companysinsertReqDto) {
@@ -112,12 +116,20 @@ public class CompanysService {
         return subcribesDao.findById(id);
     }
 
-    public void insertSubcribes(SubcribesInsertReqDto insertReqDto) {
-        // if 파인드 어쩌고해서 있는지 체크하기
+    public SubcribesInsertRespDto insertSubcribes(SubcribesInsertReqDto insertReqDto) {
+        if (companysDao.findByIdCompanyId(insertReqDto.getCompanyId()) == null) {
+            throw new RuntimeException("해당 id(" + insertReqDto.getCompanyId() + ")의 기업이 없습니다.");
+        }
         subcribesDao.insert(insertReqDto);
+        return SubcribesInsertRespDto.builder().userId(insertReqDto.getUserId()).companyId(insertReqDto.getCompanyId())
+                .build();
     }
 
+    @Transactional(rollbackFor = RuntimeException.class)
     public void deleteSubcribes(Integer id) {
+        if (subcribesDao.findById(id) == null) {
+            throw new RuntimeException("해당 기업을 구독하지 않았습니다.");
+        }
         subcribesDao.deleteById(id);
     }
